Add files via upload
This is my initial commit of a project I was working on ~11/2023. I built a full stack webmail application. This project utilized Node for the server side and React to build the client side. Functionality includes ability to add, delete contacts, view inbox / messages, as well as send and receive emails utilizing IMAP and SMTP protocol.